pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cffi==1.17.1",
    "cryptography==45.0.2",
    "pycparser==2.22",
    "livekit-agents[deepgram,elevenlabs,openai,silero,speechmatics,turn-detector,anthropic,google,azure]>=1.1.6",
    "livekit-plugins-noise-cancellation>=0.2.4",
    "protobuf==4.25.0",
    "python-dotenv"

]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "C90"]
ignore = ["E203"]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]

[tool.pylint.'MESSAGES CONTROL']
disable = ["fixme","broad-except","logging-fstring-interpolation"]
